requirements.txt
certifi==2025.1.31
Flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0

